Add single-conversion executable
