+++ CMakeLists.txt
@@ -11,10 +11,15 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -static-libgcc -static-libstdc++ -static
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-add_executable(EntryPoint entrypoint.cpp)
-
 add_subdirectory("${PROJECT_SOURCE_DIR}/IccProfLib")
-target_link_libraries(EntryPoint PUBLIC IccProfLib)
-
 add_subdirectory("${PROJECT_SOURCE_DIR}/Utils")
-target_link_libraries(EntryPoint PUBLIC Utils)
+
+# Create ConvertBulk exe and link all required code
+add_executable(ConvertBulk ConvertBulk.cpp)
+target_link_libraries(ConvertBulk PUBLIC IccProfLib)
+target_link_libraries(ConvertBulk PUBLIC Utils)
+
+# Create Convert exe and link all required code
+add_executable(ConvertSingle ConvertSingle.cpp)
+target_link_libraries(ConvertSingle PUBLIC IccProfLib)
+target_link_libraries(ConvertSingle PUBLIC Utils)
